+++ x/delayedack/keeper/rollapp_packet.go
@@ -1,13 +1,8 @@
 package keeper
 
 import (
-	errorsmod "cosmossdk.io/errors"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	transfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
-	channeltypes "github.com/cosmos/ibc-go/v6/modules/core/04-channel/types"
-	rollapptypes "github.com/dymensionxyz/dymension/v3/x/rollapp/types"
-
-	"github.com/dymensionxyz/dymension/v3/utils/gerr"
 	commontypes "github.com/dymensionxyz/dymension/v3/x/common/types"
 	"github.com/dymensionxyz/dymension/v3/x/delayedack/types"
 )
@@ -174,47 +169,4 @@ func (k Keeper) deleteRollappPacket(ctx sdk.Context, rollappPacket *commontypes.
 	}
 
 	return nil
-}
-
-// GetValidTransferWithFinalizationInfo does GetValidTransferFromReceivedPacket, but additionally it gets the finalization status and proof height
-// of the packet.
-func (k Keeper) GetValidTransferWithFinalizationInfo(
-	ctx sdk.Context,
-	packet channeltypes.Packet,
-	packetType commontypes.RollappPacket_Type,
-) (data types.TransferDataWithFinalization, err error) {
-	switch packetType {
-	case commontypes.RollappPacket_ON_RECV:
-		data.TransferData, err = k.rollappKeeper.GetValidTransfer(ctx, packet.GetData(), packet.GetDestPort(), packet.GetDestChannel())
-	case commontypes.RollappPacket_ON_TIMEOUT, commontypes.RollappPacket_ON_ACK:
-		data.TransferData, err = k.rollappKeeper.GetValidTransfer(ctx, packet.GetData(), packet.GetSourcePort(), packet.GetSourceChannel())
-	}
-	if err != nil {
-		err = errorsmod.Wrap(err, "get valid transfer data")
-		return
-	}
-
-	packetId := commontypes.NewPacketUID(packetType, packet.DestinationPort, packet.DestinationChannel, packet.Sequence)
-	height, ok := types.PacketProofHeightFromCtx(ctx, packetId)
-	if !ok {
-		// TODO: should probably be a panic
-		err = errorsmod.Wrapf(gerr.ErrNotFound, "get proof height from context: packetID: %s", packetId)
-		return
-	}
-	data.ProofHeight = height.RevisionHeight
-
-	if !data.IsRollapp() {
-		return
-	}
-
-	finalizedHeight, err := k.getRollappFinalizedHeight(ctx, data.Rollapp.RollappId)
-	if errorsmod.IsOf(err, rollapptypes.ErrNoFinalizedStateYetForRollapp) {
-		err = nil
-	} else if err != nil {
-		err = errorsmod.Wrap(err, "get rollapp finalized height")
-	} else {
-		data.Finalized = data.ProofHeight <= finalizedHeight
-	}
-
-	return
 }
